+++ src/hooks/usePlaylistRN.ts
@@ -2,9 +2,9 @@ import Snackbar from 'react-native-snackbar';
 import { useTranslation } from 'react-i18next';
 
 import { useNoxSetting } from '@stores/useApp';
-import usePlaylist from './usePlaylist';
+import usePlaylist from './usePlaylistOps';
 import useAlert from '@components/dialogs/useAlert';
-import usePlaylistOperation from './usePlaylistOperation';
+import usePlaylistAA from './usePlaylistAA';
 
 interface Props {
   callback?: () => void;
@@ -20,7 +20,7 @@ export default ({ callback = () => {} }: Props) => {
     playlistClear,
     playlistSync2Bilibili,
   } = usePlaylist();
-  const { removePlaylist } = usePlaylistOperation();
+  const { removePlaylist } = usePlaylistAA();
   const { OneWayAlert, TwoWayAlert } = useAlert();
 
   const playlistSync2BilibiliRN = async (playlist = currentPlaylist) => {
